% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_restscore.R
\name{run_boot_restscore_sequential}
\alias{run_boot_restscore_sequential}
\title{Run item-restscore bootstrap iterations sequentially}
\usage{
run_boot_restscore_sequential(
  iterations,
  boot_seeds,
  boot_data_list,
  verbose = FALSE
)
}
\arguments{
\item{iterations}{Number of iterations.}

\item{boot_seeds}{Integer vector of per-iteration seeds.}

\item{boot_data_list}{List of data passed to each worker.}

\item{verbose}{Show progress bar.}
}
\value{
List of raw results (one element per iteration).
}
\description{
Run item-restscore bootstrap iterations sequentially
}
\keyword{internal}
