% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_dependence.R
\name{run_q3_sim_sequential}
\alias{run_q3_sim_sequential}
\title{Run Q3 simulations sequentially}
\usage{
run_q3_sim_sequential(iterations, sim_seeds, sim_data_list, verbose = FALSE)
}
\arguments{
\item{iterations}{Number of iterations.}

\item{sim_seeds}{Integer vector of per-iteration seeds.}

\item{sim_data_list}{List of data passed to each worker.}

\item{verbose}{Show progress bar.}
}
\value{
List of raw results (one element per iteration).
}
\description{
Run Q3 simulations sequentially
}
\keyword{internal}
