% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_restscore_cutoff.R
\name{run_restscore_sim_parallel}
\alias{run_restscore_sim_parallel}
\title{Run item-restscore simulations in parallel using mirai}
\usage{
run_restscore_sim_parallel(
  iterations,
  sim_seeds,
  sim_data_list,
  n_cores,
  verbose = FALSE
)
}
\arguments{
\item{iterations}{Number of iterations.}

\item{sim_seeds}{Integer vector of per-iteration seeds.}

\item{sim_data_list}{List of data passed to each worker.}

\item{n_cores}{Number of mirai daemons.}

\item{verbose}{Show progress bar.}
}
\value{
List of raw results (one element per iteration).
}
\description{
Run item-restscore simulations in parallel using mirai
}
\keyword{internal}
