% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRasch2-package.R
\docType{package}
\name{easyRasch2-package}
\alias{easyRasch2}
\alias{easyRasch2-package}
\title{easyRasch2: Rasch Measurement Theory Analysis for R}
\description{
Provides functions to simplify Rasch measurement theory analysis workflows, from local dependence assessment to item fit, person fit, differential item functioning, targeting, and reliability. Uses Conditional Maximum Likelihood (CML) estimation via 'eRm' and 'psychotools' for item parameters and Weighted Likelihood Estimation (WLE) for person parameters, with 'mirt' for residual correlation analysis. Designed for easy integration with 'Quarto' documents using 'knitr::kable()' table output and 'ggplot2' figures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pgmj/easyRasch2}
  \item \url{https://pgmj.github.io/easyRasch2/}
  \item Report bugs at \url{https://github.com/pgmj/easyRasch2/issues}
}

}
\author{
\strong{Maintainer}: Magnus Johansson \email{pgmj@pm.me} (\href{https://orcid.org/0000-0003-1669-592X}{ORCID})

}
\keyword{internal}
